Change parameter of CopyFD() to uinptr, which is the type returned by os.File.Fd(). https://github.com/hughe https://github.com/hughe/gomdb/commit/cb2f62254ebe8f1ef65759978bff212e6d1b6218 and minor change

+++ env.go
@@ -132,7 +132,7 @@ func (env *Env) Copy(path string) error {
 	return errno(ret)
 }
 
-func (env *Env) CopyFd(fd int) error {
+func (env *Env) CopyFd(fd uintptr) error {
 	ret := C.mdb_env_copyfd(env._env, C.mdb_filehandle_t(fd))
 	if ret != SUCCESS {
 		return Errno(ret)
